backend/requirements.txt
beautifulsoup4==4.12.3
flask==3.0.3
flask-cors==5.0.0
Flask-SQLAlchemy==3.1.1
google-generativeai==0.8.2
gtts==2.5.3
numpy==2.1.2
pymupdf==1.24.11
python-dotenv==1.0.1
python-sonic==0.4.3
sounddevice==0.5.0